Refactor transaction handling and enhance message archival process
- Introduced `archiveBuyerMessageInbox` service to manage archival of buyer messages upon transaction actions, improving message handling consistency.
- Updated transaction-related buttons (Accept, Dispute, Reject, Resolve) to utilize the new archival service, ensuring that unread messages are archived correctly after user actions.
- Removed `TransactionDetailInvalidate` component to streamline transaction detail management, enhancing performance and reducing complexity.
- Added optional `onPostMutation` hooks to `TransactionChat` for better integration with message archival after sending messages.
- Updated README documentation to reflect changes in transaction handling and new service integration.
