+++ apps/app/src/app/@seller/README.md
@@ -34,7 +34,7 @@ Maps to server API: `/api/seller/*`.
   - `@seller/transaction/TransactionDetailPage/TransactionDetailPendingPage.tsx` is the route-level pending shell
 - Seller listing aggregate rows now use Inbox-driven unread counts grouped by listing `reference`; badge and unread emphasis must come from Inbox rather than local transaction counters.
 - Seller listing-detail transaction rows now also surface Inbox-driven unread state per transaction, using unread `buyer-message` rows whose `reference[]` contains the transaction id.
-- Opening a seller transaction detail now bulk-archives matching unread inbox rows for that transaction `reference`, so unread badges clear from both transaction and listing aggregates without requiring an extra Inbox tap.
+- Seller transaction detail no longer archives unread buyer-message inbox rows on open; seller unread state now clears only after an explicit seller action (message send, accept/reject, dispute/resolve), so browsing the thread does not prematurely mark work as handled.
 - Seller top-level transaction rows now render listing title, latest activity label, and latest activity time from the `transaction-listing` aggregate.
 - Draft editor localizes single-use read-only value components:
   - `@seller/draft/DraftEditPage/DraftEditor/value/CategoryValue.tsx`
