+++ apps/app/src/app/@seller/transaction/ui/button/RejectButton.tsx
@@ -1,3 +1,4 @@
+import { useQueryClient } from "@tanstack/react-query";
 import { ConfirmButton } from "@use-pico/client/ui/button";
 import { Tx } from "@use-pico/client/ui/tx";
 import type { tTransaction } from "@zbav-se.me/sdk/api/seller";
@@ -5,6 +6,7 @@ import { withTransactionRejectMutation } from "@zbav-se.me/sdk/mutation/seller/t
 import { CancelIcon } from "@zbav-se.me/ui/icon";
 import type { FC } from "react";
 import type { TransactionMenuButton } from "~/app/@common/transaction/ui/TransactionMenuButton";
+import { archiveBuyerMessageInbox } from "../../service/archiveBuyerMessageInbox";
 
 export namespace RejectButton {
 	export interface Props extends Partial<ConfirmButton.Props> {
@@ -14,8 +16,19 @@ export namespace RejectButton {
 }
 
 export const RejectButton: FC<RejectButton.Props> = ({ close, transaction, ...props }) => {
+	const queryClient = useQueryClient();
 	const mutation = withTransactionRejectMutation.useMutation({
-		onSuccess() {
+		async onPostMutation() {
+			try {
+				await archiveBuyerMessageInbox({
+					queryClient,
+					transactionId: transaction.id,
+					listingId: transaction.listingId,
+				});
+			} catch {
+				// Keep reject flow usable even if unread archival fails.
+			}
+
 			close();
 		},
 	});
