+++ apps/app/src/app/@seller/transaction/ui/button/ResolveButton.tsx
@@ -1,3 +1,4 @@
+import { useQueryClient } from "@tanstack/react-query";
 import { Button } from "@use-pico/client/ui/button";
 import { Tx } from "@use-pico/client/ui/tx";
 import type { tTransaction } from "@zbav-se.me/sdk/api/seller";
@@ -5,6 +6,7 @@ import { withTransactionResolveMutation } from "@zbav-se.me/sdk/mutation/seller/
 import { CheckIcon } from "@zbav-se.me/ui/icon";
 import type { FC } from "react";
 import type { TransactionMenuButton } from "~/app/@common/transaction/ui/TransactionMenuButton";
+import { archiveBuyerMessageInbox } from "../../service/archiveBuyerMessageInbox";
 
 export namespace ResolveButton {
 	export interface Props extends Button.Props {
@@ -14,8 +16,19 @@ export namespace ResolveButton {
 }
 
 export const ResolveButton: FC<ResolveButton.Props> = ({ close, transaction, ...props }) => {
+	const queryClient = useQueryClient();
 	const mutation = withTransactionResolveMutation.useMutation({
-		onSuccess() {
+		async onPostMutation() {
+			try {
+				await archiveBuyerMessageInbox({
+					queryClient,
+					transactionId: transaction.id,
+					listingId: transaction.listingId,
+				});
+			} catch {
+				// Keep resolve flow usable even if unread archival fails.
+			}
+
 			close();
 		},
 	});
